Web Browser App: Minor refactoring
